Restored the last few changes from the rename and file move.
I had started editing those files for another TODO item, and it was more involved code changes than I wanted mixed in with file renames/moves, so I created a new branch for those changed files and forgot that the import paths in them had changed.  I restored those path changes.

Files checked in: DataRepo/utils/compounds_loader.py DataRepo/utils/protocols_loader.py DataRepo/utils/sequences_loader.py DataRepo/utils/study_table_loader.py DataRepo/utils/tissues_loader.py
