+++ DataRepo/utils/study_table_loader.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 
 from DataRepo.models import Study
-from DataRepo.utils.loader import TraceBaseLoader
+from DataRepo.utils.table_loader import TraceBaseLoader
 
 
 class StudyTableLoader(TraceBaseLoader):
