+++ DataRepo/utils/tissues_loader.py
@@ -1,7 +1,7 @@
 from collections import namedtuple
 
 from DataRepo.models import Tissue
-from DataRepo.utils.loader import TraceBaseLoader
+from DataRepo.utils.table_loader import TraceBaseLoader
 
 
 class TissuesLoader(TraceBaseLoader):
